Change the file open mode in append function
Open in file create function is done in binary mode (set by "b" flag in
the second argument of open()), and this was not done the same way in
append function. Change that, and also enclode a line that we are
appending, since open is in binary mode now, and all strings need to be
encoded.