Change stopAllTasks to return instead of yielding
stopAllTasks in TaskManager.py attempts to block its thread and thus
yield control to manager thread for that one to terminate itself, but
this is not happening unless stopAllTasks returns, causing the whole
program to freeze, since Tkinter thread hangs.

This fix is questionable, as we get rid of synchoronization logic that
waits for manager to terminate before proceeding with other tasks. But
for not, it is thoughs and prayers.