Redo stdout redirect to use custom class
Previously I implemented stdout redirection using decorator functions to
duplicate stdout both to console and GUI window. This method works fine
with a script, but stops working when script is compiled to .exe,
because pyinstaller overwrites stdout.write with some kind of null
object. This implementation not just decorates the function
stdout.writes, but replaces the whole class of stdout with a custom one,
that handles writes by outputting them to GUI window as needed.