Fix bacnet reader not finidng devices bug
Unhandled exception occurs when trying to find devices on IP and port
that is already being used by other application. Handle this properly
and display error message.

This commit also fixes a bug where no IPs will be shown when none of the
adapters have 'Ethernet' in the name. Handle this case by displaying all
IPs in such case regardless.