Error handling, changes to status messages, WhoIs
Status messages on the GUI are more verbose now.

Add more error handling for some of the functions.

Add additional WhoIs request after bacnet.discover. This is a fix for a
bug where controller 100 would not get discovered by BAC0, but all the
devices on all the networks that 100 routes to would, if controller 100
has its Ethernet adapter disabled.