Clear object list on network switch. Lock buttons.
Objects were being preserved in the list even when different network was
selected and device that those objects got loaded from is not on this
network and no longer showing up in the drop down. This is logically
incorrect. Fix this.

Some undesired side effects are happening when trying to switch network
while objects are loading. Prevent user from being able to do that.